% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_predictors.R
\name{compute_caa_ra}
\alias{compute_caa_ra}
\title{Compute the CAA or CAA score of two nodes}
\usage{
compute_caa_ra(ne, i, j, type = "CAA")
}
\arguments{
\item{ne}{list; The neighbourhood of all nodes from a network of interest.}

\item{i}{integer; The ID of a network node.}

\item{j}{integer; The ID of a network node.}

\item{type}{character; One of "CAA" or "CRA".}
}
\value{
The CAA or CRA likelihood score of interaction for the node pair i-j.
}
\description{
Given two indices representing two nodes and a list of neighbourhoods,
compute the CAA or CRA index for the node pair. This function is used by 
\code{lp_caa} and \code{lp_cra}.
}
\references{
Cannistraci, C. V. et al. (2013) From link-prediction in brain 
connectomes and protein interactomes to the local-community-paradigm in 
complex networks. \emph{Sci. Rep.} 3:1613
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
