% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_predictors.R
\name{compute_ra}
\alias{compute_ra}
\title{Compute the Resource Allocation score of two nodes}
\usage{
compute_ra(ne, i, j)
}
\arguments{
\item{ne}{list; The neighbourhood of all nodes from a network of interest.}

\item{i}{integer; The ID of a network node.}

\item{j}{integer; The ID of a network node.}
}
\value{
The Resource Allocation likelihood score of interaction for the node
pair i-j.
}
\description{
Given two indices representing two nodes and a list of neighbourhoods,
compute the Resource Allocation index for the node pair. This function is 
used by \code{lp_ra}.
}
\references{
Zhou, T. et al. (2009) Predicting missing links via local 
information. \emph{The European Physical Journal B} 71(4):623-630
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
