% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_predictors.R
\name{get_non_edges}
\alias{get_non_edges}
\title{Disconnected node pairs in a network}
\usage{
get_non_edges(g)
}
\arguments{
\item{g}{igraph; The network of interest.}
}
\value{
A matrix with the disconnected node pairs in the network.
}
\description{
Given a network of interest with N nodes and L links, produces a matrix with
N(N-1)/2 - L rows and 2 columns containing the list of disconnected node
pairs in the network.
}
\examples{
# Determine which node pairs are disconnected in the Zacharay Karate Club 
# network
non_edges <- get_non_edges(g = karate_club)

}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
