% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karate_club_data.R
\docType{data}
\name{karate_club}
\alias{karate_club}
\title{Ties between members of a karate club}
\format{An igraph object representing the network}
\source{
\href{http://konect.uni-koblenz.de/networks/ucidata-zachary}{Koblenz 
Network Collection.}
}
\usage{
karate_club
}
\description{
Ties between the members of a karate club as recorded by Wayne Zachary in 
1977.
}
\examples{
# Show number of nodes and edges
igraph::vcount(karate_club)
igraph::ecount(karate_club)

}
\references{
Zachary, Wayne (1977) An information flow model for conflict and 
fission in small groups. \emph{J. of Anthropological Research} 33:452-473
}
\keyword{datasets}
\keyword{igraph}
\keyword{karate}
\keyword{zachary}
