% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkprediction-package.r
\docType{package}
\name{linkprediction}
\alias{linkprediction}
\alias{linkprediction-package}
\title{Link Prediction Methods}
\description{
Implements most of existing methods proximity-based methods of link
prediction in graphs. See \code{\link{proxfun}}.
}
\note{
Authors thank (Polish) National Science Centre for support through SONATA 
grant 2012/07/D/HS6/01971 for the project \emph{Dynamics of Competition and
Collaboration  in Science: Individual Strategies, Collaboration Networks, and
Organizational Hierarchies} (\url{recon.icm.edu.pl}).
}
