% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_predictors.R
\name{lp_aa}
\alias{lp_aa}
\title{Link prediction with Adamic and Adar's index}
\usage{
lp_aa(g)
}
\arguments{
\item{g}{igraph; The network of interest.}
}
\value{
Tibble with the following columns:
\item{nodeA}{The ID of a network node.}
\item{nodeB}{The ID of a network node.}
\item{scr}{The likelihood score of interaction for the node pair.}
}
\description{
Given a network of interest, it computes the likelihood score of interaction,
for all disconnected node pairs, with the Adamic and Adar link predictor.
}
\examples{
# Apply the Adamic and Adar link predictor to the Zachary Karate Club network
aa <- lp_aa(g = karate_club)

}
\references{
Adamic, L. and Adar, E. (2003) Friends and neighbors on the Web. 
\emph{Social Networks} 25(3):211-230
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
