% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_predictors.R
\name{lp_cn}
\alias{lp_cn}
\title{Link prediction with Common Neighbours}
\usage{
lp_cn(g)
}
\arguments{
\item{g}{igraph; The network of interest.}
}
\value{
Tibble with the following columns:
\item{nodeA}{The ID of a network node.}
\item{nodeB}{The ID of a network node.}
\item{scr}{The likelihood score of interaction for the node pair.}
}
\description{
Given a network of interest, it computes the likelihood score of interaction,
for all disconnected node pairs, with the Common Neighbours link predictor.
}
\examples{
# Apply the Common Neighbours link predictor to the Zachary Karate Club 
# network
cn <- lp_cn(g = karate_club)

}
\references{
Newman, M. E. J. (2001) Clustering and preferential attachment in
growing networks. \emph{Phys. Rev. E.} 64:025102
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
