% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_predictors.R
\name{lp_dice}
\alias{lp_dice}
\title{Link prediction with Dice's index}
\usage{
lp_dice(g)
}
\arguments{
\item{g}{igraph; The network of interest.}
}
\value{
Tibble with the following columns:
\item{nodeA}{The ID of a network node.}
\item{nodeB}{The ID of a network node.}
\item{scr}{The likelihood score of interaction for the node pair.}
}
\description{
Given a network of interest, it computes the likelihood score of interaction,
for all disconnected node pairs, with the Dice link predictor.
}
\examples{
# Apply the Dice link predictor to the Zachary Karate Club network
dice <- lp_dice(g = karate_club)

}
\references{
Dice, L. R. (1945) Measures of the amount of ecologic 
association between species. \emph{Ecology} 26(3):297-302
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
