% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_predictors.R
\name{lp_igraph}
\alias{lp_igraph}
\title{Link prediction with methods available in igraph}
\usage{
lp_igraph(g, method)
}
\arguments{
\item{g}{igraph; The network of interest.}

\item{method}{character; One of "jaccard", "dice" or "invlogweighted".}
}
\value{
Tibble with the following columns:
\item{nodeA}{The ID of a network node.}
\item{nodeB}{The ID of a network node.}
\item{scr}{The likelihood score of interaction for the node pair.}
}
\description{
Given a network of interest, it computes the likelihood score of interaction,
for all disconnected node pairs, with the link predictors provided by igraph.
This function is used by \code{lp_jc}, \code{lp_dice} and \code{lp_aa}.
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
