% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_predictors.R
\name{lp_isomap}
\alias{lp_isomap}
\title{Link prediction with ISOMAP}
\usage{
lp_isomap(g, d = 2, use_weights = FALSE)
}
\arguments{
\item{g}{igraph; The network of interest.}

\item{d}{integer; The dimension of the embedding space.}

\item{use_weights}{logical; Indicates whether edge weights should be used to
compute shortest paths between nodes.}
}
\value{
Tibble with the following columns:
\item{nodeA}{The ID of a network node.}
\item{nodeB}{The ID of a network node.}
\item{scr}{The likelihood score of interaction for the node pair.}
}
\description{
Given a network of interest, it computes the likelihood score of interaction,
for all disconnected node pairs, based on the embedding of the network to 
d-dimensional Euclidean space with ISOMAP.
}
\examples{
# Apply the ISOMAP link predictor to the Zachary Karate Club network
iso <- lp_isomap(g = karate_club, d = 2, use_weights = FALSE)

}
\references{
Tenenbaum, J. B. (2000)  A global geometric framework for 
nonlinear dimensionality reduction. \emph{Science} 290:2319-2323

Kuchaiev, O. et al. (2009)  Geometric de-noising of 
protein-protein interaction networks. \emph{PLoS Comput. Biol.} 5:e1000454
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
