% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_predictors.R
\name{lp_jc}
\alias{lp_jc}
\title{Link prediction with Jaccard's index}
\usage{
lp_jc(g)
}
\arguments{
\item{g}{igraph; The network of interest.}
}
\value{
Tibble with the following columns:
\item{nodeA}{The ID of a network node.}
\item{nodeB}{The ID of a network node.}
\item{scr}{The likelihood score of interaction for the node pair.}
}
\description{
Given a network of interest, it computes the likelihood score of interaction,
for all disconnected node pairs, with the Jaccard link predictor.
}
\examples{
# Apply the Jaccard link predictor to the Zachary Karate Club network
jc <- lp_jc(g = karate_club)

}
\references{
Jaccard, P. (1912) The distribution of the flora in the alpine 
zone. \emph{New phytologist} 11(2):37-50
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
