% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_predictors.R
\name{lp_l3}
\alias{lp_l3}
\title{Link prediction with L3}
\usage{
lp_l3(g)
}
\arguments{
\item{g}{igraph; The network of interest.}
}
\value{
Tibble with the following columns:
\item{nodeA}{The ID of a network node.}
\item{nodeB}{The ID of a network node.}
\item{scr}{The likelihood score of interaction for the node pair.}
}
\description{
Given a network of interest, it computes the likelihood score of interaction,
for all disconnected node pairs, with the L3 link predictor.
}
\examples{
# Apply the L3 link predictor to the Zachary Karate Club network
l3 <- lp_l3(g = karate_club)

}
\references{
Kovacs, I. A., et al. (2018) Network-based prediction of protein
interactions. \emph{bioRxiv 275529} doi: https://doi.org/10.1101/275529
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
