% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_predictors.R
\name{lp_leig}
\alias{lp_leig}
\title{Link prediction with Laplacian Eigenmaps}
\usage{
lp_leig(g, d = 2, use_weights = FALSE)
}
\arguments{
\item{g}{igraph; The network of interest.}

\item{d}{integer; The dimension of the embedding space.}

\item{use_weights}{logical; Indicates whether edge weights should be used to
compute the network's Laplacian matrix.}
}
\value{
Tibble with the following columns:
\item{nodeA}{The ID of a network node.}
\item{nodeB}{The ID of a network node.}
\item{scr}{The likelihood score of interaction for the node pair.}
}
\description{
Given a network of interest, it computes the likelihood score of interaction,
for all disconnected node pairs, based on the embedding of the network to 
d-dimensional Euclidean space with Laplacian Eigenmaps.
}
\examples{
# Apply the Laplacian Eigenmaps link predictor to the Zachary Karate Club 
# network
leig <- lp_leig(g = karate_club, d = 2, use_weights = FALSE)

}
\references{
Belkin, M. and Niyogi, P. (2001) Laplacian eigenmaps and 
spectral techniques for embedding and clustering. \emph{Adv. Neur. In.} 
14:585-591
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
