% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_predictors.R
\name{lp_matrix}
\alias{lp_matrix}
\title{Link prediction with pairwise node (dis-)similarities}
\usage{
lp_matrix(g, m, type = "sim")
}
\arguments{
\item{g}{igraph; The network of interest.}

\item{m}{matrix; A square matrix of pairwise node (dis-)similarities.}

\item{type}{character; One of "sim" or "dis", which indicate whether the 
matrix stores similarities (high values -> higher likelihood of interaction) 
or dissimilarities (low values -> higher likelihood of interaction).}
}
\value{
Tibble with the following columns:
\item{nodeA}{The ID of a network node.}
\item{nodeB}{The ID of a network node.}
\item{scr}{The likelihood score of interaction for the node pair.}
}
\description{
Given a network of interest and a matrix of pairwise node (dis-)similarities
produced by some algorithm, it computes the likelihood score of interaction,
for all disconnected node pairs, based on such (dis-)similarities.
}
\examples{
# Compute a node similarity matrix with one of the algorithms included in 
# igraph
dice <- igraph::similarity(karate_club, method = "dice")

# Now compute likelihood scores for disconnected node pairs with this matrix
sim <- lp_matrix(g = karate_club, dice, type = "sim")

}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
