% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_predictors.R
\name{lp_mce}
\alias{lp_mce}
\title{Link prediction with MCE or ncMCE}
\usage{
lp_mce(g, d = 2, centre = FALSE, use_weights = FALSE)
}
\arguments{
\item{g}{igraph; The network of interest.}

\item{d}{integer; The dimension of the embedding space.}

\item{centre}{logical; Indicates whether the Minimum Curvilinear kernel 
should be centred or not.}

\item{use_weights}{logical; Indicates whether edge weights should be used to
compute the network's Minimum Curvilinear kernel.}
}
\value{
Tibble with the following columns:
\item{nodeA}{The ID of a network node.}
\item{nodeB}{The ID of a network node.}
\item{scr}{The likelihood score of interaction for the node pair.}
}
\description{
Given a network of interest, it computes the likelihood score of interaction,
for all disconnected node pairs, based on the embedding of the network to 
d-dimensional Euclidean space with centred or non-centred Minimum Curvilinear
Embedding (MCE and ncMCE, respectively).
}
\examples{
# Apply the ncMCE link predictor to the Zachary Karate Club 
# network
mce <- lp_mce(g = karate_club, d = 2, centre = FALSE, use_weights = FALSE)

}
\references{
Cannistraci, C. V. et al. (2013) Minimum curvilinearity to 
enhance topological prediction of protein interactions by network embedding.
\emph{Bioinformatics} 29:i199-i209
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
