% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_predictors.R
\name{lp_ra}
\alias{lp_ra}
\title{Link prediction with Resource Allocation}
\usage{
lp_ra(g)
}
\arguments{
\item{g}{igraph; The network of interest.}
}
\value{
Tibble with the following columns:
\item{nodeA}{The ID of a network node.}
\item{nodeB}{The ID of a network node.}
\item{scr}{The likelihood score of interaction for the node pair.}
}
\description{
Given a network of interest, it computes the likelihood score of interaction,
for all disconnected node pairs, with the Resource Allocation link predictor.
}
\examples{
# Apply the Resource Allocation link predictor to the Zachary Karate Club 
# network
ra <- lp_ra(g = karate_club)

}
\references{
Zhou, T. et al. (2009) Predicting missing links via local 
information. \emph{The European Physical Journal B} 71(4):623-630
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
