% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_predictors.R
\name{lp_spm}
\alias{lp_spm}
\title{Link prediction with the SPM}
\usage{
lp_spm(g, p_H = 0.1, epochs = 10, k = NA)
}
\arguments{
\item{g}{igraph; The network of interest.}

\item{p_H}{numeric; Fraction of network links to remove (perturbation set).}

\item{epochs}{integer; Number of perturbation sets to consider.}

\item{k}{integer; If k != NA and k < N (number of network nodes), the method 
is approximated by computing the k larget eigenvalues instead of all N.}
}
\value{
Tibble with the following columns:
\item{nodeA}{The ID of a network node.}
\item{nodeB}{The ID of a network node.}
\item{scr}{The likelihood score of interaction for the node pair.}
}
\description{
Given a network of interest, it computes the likelihood score of interaction,
for all disconnected node pairs, with the Structural Perturbation Method 
(SPM).
}
\examples{
# Apply the approximated SPM link predictor to the Zachary Karate Club 
# network
spm <- lp_spm(g = karate_club, p_H = 0.1, epochs = 10, 
              k = round(sqrt(igraph::vcount(karate_club))))

}
\references{
Lu, L. et al. (2015) Toward link predictability of complex 
networks. \emph{PNAS} 112(8):2325-2330
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
