% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_link_predictors.R
\name{plot_lp_precision}
\alias{plot_lp_precision}
\title{Plot the performance evaluation of link predictors}
\usage{
plot_lp_precision(res, colours = NA)
}
\arguments{
\item{res}{tibble; The result of applying \code{prune_recover} to a network.}

\item{colours}{character/numeric; A vector of colours to depict the 
performance curve of each link prediction method. There should be as many
colours as assessed link predictors.}
}
\value{
A ggplot with the precision curve(s) of the assesed prediction 
method(s).
}
\description{
Given the assessment of one or more link predictors with function 
\code{prune_recover}, plots the precision curves of the analysed methods.
Each point corresponds to the average precision for a specific fraction of
removed links across all considered epochs.
}
\examples{
# Assess the performance of three link predictors applied to the Zachary 
# Karate Club network
assessment <- prune_recover(g = karate_club, "lp_cn", "lp_aa", "lp_pa")

# Define a set of colours to plot
colours <- c("#8da0cb", "#fc8d62", "#66c2a5")

# Plot the precision curves of the considered link predictors
perf <- plot_lp_precision(assessment, colours)

}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
