% Generated by roxygen2 (4.0.2): do not edit by hand
\name{proxfun}
\alias{proxfun}
\title{Node proximity}
\usage{
proxfun(graph, method, v1 = NULL, v2 = v1, ...)
}
\arguments{
\item{graph}{an object of class igraph}

\item{method}{a method (single string) for calculating similarities, see Details}

\item{v1,v2}{vectors of vertices between which similarity will be calculated
character is treated as names, numeric as ids}

\item{...}{additional arguments specific for a selected method}
}
\value{
matrix
}
\description{
General function for calculating nodes' proximity in a graph
}
\details{
This function calculates vertex proximity (similarity) with selected method
  and between selected vertices.

Following methods are available:

 \code{act} average commute time

 \code{act_n} normalized average commute time

 \code{aa} Adamic-Adar index

 \code{cn} common neighbours

 \code{cos} cosine similarity

 \code{cos_l} cosine similarity on L+

 \code{dist} graph distance

 \code{hdi} Hub Depressed Index

 \code{hpi} Hub Promoted Index

 \code{jaccard} Jaccard coefficient

 \code{katz} Katz index

 \code{l} L+ directly

 \code{lhn_local} Leicht-Holme-Newman Index

 \code{lhn_global} Leicht-Holme-Newman Index global version

 \code{lp} Local Path Index

 \code{mf} Matrix Forest Index

 \code{pa} preferential attachment

 \code{ra} resource allocation

 \code{rwr} random walk with restart

 \code{sor} sorensen index/ dice coefficient
}

