% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_aa}
\alias{similarity_aa}
\title{Adamic-Adar index}
\usage{
similarity_aa(graph)
}
\arguments{
\item{graph}{an object of class igraph}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph.
}
\description{
Similarity measure counting common neighbors weighted by their degrees.
}
\details{
Adamic-Adar Index counts common neighbors with weigths equal to logarithms of
  neighbors' degrees. This is a wrapper to an \pkg{igraph} function
  \code{\link[igraph]{similarity.invlogweighted}}.
}
\examples{
g <- igraph::random.graph.game(20, 0.3)
similarity_aa(g)
}
\seealso{
\code{\link[igraph]{similarity}}
}

