% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_act}
\alias{similarity_act}
\title{Averag Commute Time}
\usage{
similarity_act(graph)
}
\arguments{
\item{graph}{an object of class igraph}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph.
}
\description{
Similarity measure based on random walker
}
\details{
This function calculates vertex similarity based on the average number of
 steps, that random walker on the graph needs to get from one vertex to
 another.
}
\examples{
g <- igraph::random.graph.game(20, 0.3)
similarity_act(g)
}
\seealso{
\code{\link[igraph]{similarity}}
}

