% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_act_normal}
\alias{similarity_act_normal}
\title{Averag Commute Time, normalized}
\usage{
similarity_act_normal(graph)
}
\arguments{
\item{graph}{an object of class igraph}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph.
}
\description{
Similarity measure based on random walker
}
\details{
This function calculates vertex similarity based on the average number of
 steps, that random walker on the graph needs to get from one vertex to
 another and additionally weights it by its stationary distribution.
}
\examples{
g <- igraph::random.graph.game(20, 0.3)
similarity_act_normal(g)
}
\seealso{
\code{\link[igraph]{similarity}}
}

