% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_cn}
\alias{similarity_cn}
\title{Common neighbors vertex similarity}
\usage{
similarity_cn(graph)
}
\arguments{
\item{graph}{an object of class igraph}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph.
}
\description{
Similarity measure counting number of common neighbors.
}
\details{
This is a simple wrapper to an \pkg{igraph} function \code{cocitation},
  which counts the number of neighbours shared by two vertices.
}
\examples{
g <- igraph::random.graph.game(20, 0.3)
similarity_cn(g)
}
\seealso{
\code{\link[igraph]{similarity}}

\code{\link[igraph]{cocitation}}
}

