% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_cos}
\alias{similarity_cos}
\title{Cosine vertex similarity/ Salton index}
\usage{
similarity_cos(graph)
}
\arguments{
\item{graph}{an object of class igraph}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph.
}
\description{
Similarity measure based on resource allocation process.
}
\details{
This function measures the cosine of the angle between columns of
  the adjacency matrix, corresponding to given nodes.
}
\examples{
g <- igraph::random.graph.game(20, 0.3)
similarity_cos(g)
}
\seealso{
\code{\link[igraph]{similarity}}
}

