% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_cosineL}
\alias{similarity_cosineL}
\title{Cosine similarity  on \eqn{L^+}}
\usage{
similarity_cosineL(graph)
}
\arguments{
\item{graph}{an object of class igraph}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph.
}
\description{
Similarity measure based on the pseudoinverse of the Laplacian matrix
}
\details{
This function measures the cosine of the angle between columns of
  the pseudoinverse of the Laplacian matrix.
}
\examples{
g <- igraph::random.graph.game(20, 0.3)
similarity_cosineL(g)
}
\seealso{
\code{\link[igraph]{similarity}}
}

