% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_distance}
\alias{similarity_distance}
\title{Geodesic distance vertex similarity}
\usage{
similarity_distance(graph)
}
\arguments{
\item{graph}{an object of class igraph}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph.
}
\description{
This function calculates similarity score for vertices based on the
  shortest paths between them.
}
\details{
Geodesic distance between vertices is the length of the shortest paths
  between them. If there is no path between vertices (ie. thay are in
  different connected components), a geodesic distance is set to the number
  of vertices in the graph plus one. Distances are inversed afterwards, so as
  two vertices are more similar when the distances between them is shorter.
}
\examples{
g <- igraph::random.graph.game(20, 0.3)
similarity_distance(g)
}
\seealso{
\code{\link[igraph]{similarity}}
}

