% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_hdi}
\alias{similarity_hdi}
\title{Hub Depressed Index}
\usage{
similarity_hdi(graph)
}
\arguments{
\item{graph}{an object of class igraph}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph.
}
\description{
Similarity measure based on common neighbors
}
\details{
This measures assigns lower scores to links adjacent to hubs (high degree
  nodes). It counts common neighbors of two vertices and weigths the result
  by the higher of those vertices' degrees.
}
\examples{
g <- igraph::random.graph.game(20, 0.3)
similarity_hdi(g)
}
\seealso{
\code{\link[igraph]{similarity}}
}

