% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_hpi}
\alias{similarity_hpi}
\title{Hub Promoted Index}
\usage{
similarity_hpi(graph)
}
\arguments{
\item{graph}{an object of class igraph}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph.
}
\description{
Similarity measure based on common neighbors
}
\details{
This measures assigns higher scores to links adjacent to hubs (high degree
  nodes). It counts common neighbors of two vertices and weigths the result
  by the lower of those vertices' degrees.
}
\examples{
g <- igraph::random.graph.game(20, 0.3)
similarity_hpi(g)
}
\seealso{
\code{\link[igraph]{similarity}}
}

