% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_jaccard}
\alias{similarity_jaccard}
\title{Jaccard index}
\usage{
similarity_jaccard(graph)
}
\arguments{
\item{graph}{an object of class igraph}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph.
}
\description{
Similarity measure counting the proportion of shared nodes.
}
\details{
This is a simple wrapper to an \pkg{igraph} function
  \code{\link[igraph]{similarity.jaccard}}, which calculates the proportion
  of shared neighbors of two vertices.
}
\examples{
g <- igraph::random.graph.game(20, 0.3)
similarity_jaccard(g)
}
\seealso{
\code{\link[igraph]{similarity}}
}

