% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_katz}
\alias{similarity_katz}
\title{Katz Index}
\usage{
similarity_katz(graph, beta = 0.001)
}
\arguments{
\item{graph}{an object of class igraph}

\item{beta}{damping factor}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph.
}
\description{
Similarity measure based on all paths in a graph
}
\details{
This function counts all the paths between given pair of nodes, with shorter
  paths counting more heavily. Weigths are exponential.
}
\examples{
g <- igraph::random.graph.game(20, 0.3)
similarity_katz(g, 0.001)
}
\seealso{
\code{\link[igraph]{similarity}}
}

