% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_l}
\alias{similarity_l}
\title{Pseudoinverse of the Laplacian}
\usage{
similarity_l(graph)
}
\arguments{
\item{graph}{an object of class igraph}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph.
}
\description{
Similarity measure based solely on the pseudoinverse of the Laplacian matrix
}
\details{
The pseudo inverse of the Laplacian matrix provides provides a direct measure
  of similarity, as its elements are the inner products of vectors from an
  Euclidean space, which preserves Average Commute Time between nodes.
}
\examples{
g <- igraph::random.graph.game(20, 0.3)
similarity_l(g)
}
\seealso{
\code{\link[igraph]{similarity}}
}

