% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_lhn_global}
\alias{similarity_lhn_global}
\title{Leicht-Holme-Newman Global Index}
\usage{
similarity_lhn_global(graph, theta = 0.5)
}
\arguments{
\item{graph}{an object of class igraph}

\item{theta}{damping factor}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph.
}
\description{
Recursive similarity measure based on all paths in a graph
}
\details{
This function calculates vertex similarity based on the concept, that two
  nodes are similar, when their neighbors are similar. Smaller \eqn{\theta}
  assings more weight on shorter paths (closer neighbors are more important).
}
\examples{
g <- igraph::random.graph.game(20, 0.3)
similarity_lhn_global(g, 0.5)
}
\seealso{
\code{\link[igraph]{similarity}}
}

