% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_lhn_local}
\alias{similarity_lhn_local}
\title{Leicht-Holme-Newman Index}
\usage{
similarity_lhn_local(graph)
}
\arguments{
\item{graph}{an object of class igraph}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph.
}
\description{
Similarity measure based on the number of common neighbors.
}
\details{
This function counts the number of common neighbours relative to the expected
  value (in random graph with the same degree distribution).
}
\examples{
g <- igraph::random.graph.game(20, 0.3)
similarity_lhn_local(g)
}
\seealso{
\code{\link[igraph]{similarity}}
}

