% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_lp}
\alias{similarity_lp}
\title{Local Paths Index}
\usage{
similarity_lp(graph, eps = 0.01)
}
\arguments{
\item{graph}{an object of class igraph}

\item{eps}{the importance of paths of length three}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph.
}
\description{
Similarity measure based on the neighborhood of order 2
}
\details{
This function counts the number of two-paths and three-paths between nodes,
  with three-paths weighted by a parameter \eqn{\epsilon}.
}
\examples{
g <- igraph::random.graph.game(20, 0.3)
similarity_lp(g, eps = 0.01)
}
\seealso{
\code{\link[igraph]{similarity}}
}

