% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_matrix_forest}
\alias{similarity_matrix_forest}
\title{Matrix Forest Index}
\usage{
similarity_matrix_forest(graph)
}
\arguments{
\item{graph}{an object of class igraph}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph.
}
\description{
Similarity measure based on the concept of spanning trees
}
\details{
This measure can be understood as the ratio of the number of spanning rooted
  forest such that nodes \eqn{x} and \eqn{y} belong to the same tree rooted
  at \eqn{x} to all spanning rooted forests of the network.
}
\examples{
g <- igraph::random.graph.game(20, 0.3)
similarity_matrix_forest(g)
}
\seealso{
\code{\link[igraph]{similarity}}
}

