% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_pa}
\alias{similarity_pa}
\title{Preferential attachment vertex similarity}
\usage{
similarity_pa(graph)
}
\arguments{
\item{graph}{an object of class igraph}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph.
}
\description{
This function calculates similarity score for vertices based on their degrees.
}
\details{
Preferential attachment model was designed by Barabsi and Albert to simulate
  the evolution of scale-free graphs. The probability of new edge occuring
  between two vertices is proportional to the product of degrees of those
  vertices. It is a similarity measure under assumption that the most similar
  vertices are simulataneously the most probable to be connected.
}
\examples{
g <- igraph::barabasi.game(20)
similarity_pa(g)
}
\seealso{
\code{\link[igraph]{similarity}}
}

