% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_ra}
\alias{similarity_ra}
\title{Resource allocation vertex similarity}
\usage{
similarity_ra(graph)
}
\arguments{
\item{graph}{an object of class igraph}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph.
}
\description{
Similarity measure based on resource allocation process.
}
\details{
This function counts the number of common neighbours weighted by the inverse
of their degrees.
}
\examples{
g <- igraph::random.graph.game(20, 0.3)
similarity_ra(g)
}
\seealso{
\code{\link[igraph]{similarity}}
}

