% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_resource_allocation}
\alias{similarity_resource_allocation}
\title{Similarity measure.}
\usage{
similarity_resource_allocation(graph)
}
\arguments{
\item{graph}{Igraph object.}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph
}
\description{
Similarity measure based on resource allocation process.
}
\details{
This function counts the number of common neighbours weighted by the inverse
of their degrees.
}
\examples{
g <- igraph::random.graph.game(20, 0.3)
similarity_resource_allocation(g)
}
\seealso{
\code{\link[devtools]{document}}
}

