% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_rwr}
\alias{similarity_rwr}
\title{Random Walk with Restart}
\usage{
similarity_rwr(graph, alpha = 0.3)
}
\arguments{
\item{graph}{an object of class igraph}

\item{alpha}{probability of return of random walker}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph.
}
\description{
Similarity measure based on random walker returning to its root
}
\details{
It is an adaptation of PageRank algorithm. This function calculates
  similarity between two nodes \eqn{a} and \eqn{b} as a sum of stationary
  distributions of one node in random walk process starting from the other
  node. Random walker could return to its root at every step with fixed
  fixed probability \eqn{\alpha}.
}
\examples{
g <- igraph::random.graph.game(20, 0.3)
similarity_rwr(g, alpha = 0.3)
}
\seealso{
\code{\link[igraph]{similarity}}
}

