% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_sorensen}
\alias{similarity_sorensen}
\title{Sorensen index/ Dice's coefficient}
\usage{
similarity_sorensen(graph)
}
\arguments{
\item{graph}{an object of class igraph}
}
\value{
A square numeric matrix with size equal to the number of vertices
  in the input graph.
}
\description{
Similarity measure counting the proportion of shared nodes.
}
\details{
This function measures a relative size of an intersection of neighbors' sets
  of two vertices. It is a wrapper of an \pkg{igraph} function
  \code{\link[igraph]{similarity.dice}}.
}
\examples{
g <- igraph::random.graph.game(20, 0.3)
similarity_sorensen(g)
}
\seealso{
\code{\link[igraph]{similarity}}
}

