% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_predictors.R
\name{structural_consistency}
\alias{structural_consistency}
\title{Structural consistency of a network}
\usage{
structural_consistency(g, p_H = 0.1, epochs = 100, k = NA)
}
\arguments{
\item{g}{igraph; The network of interest.}

\item{p_H}{numeric; Fraction of network links to remove (perturbation set).}

\item{epochs}{integer; Number of perturbation sets to consider.}

\item{k}{integer; If k != NA and k < N (number of network nodes), the method 
is approximated by computing the k larget eigenvalues instead of all N.}
}
\value{
A vector with \code{epochs} elements containing the consistency 
values for each perturbation set.
}
\description{
Given a network of interest, it computes its Structural Consistency sigma_c.
High values of sigma_c correlate with higher predictability of missing links
in a network.
}
\examples{
# Compute the structural consistency of the Zachary Karate Club network
sigma_c <- structural_consistency(g = karate_club, p_H = 0.1, epochs = 100, 
              k = round(sqrt(igraph::vcount(karate_club))))
avg_sigma_c <- mean(sigma_c)

}
\references{
Lu, L. et al. (2015) Toward link predictability of complex 
networks. \emph{PNAS} 112(8):2325-2330
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
