% Generated by roxygen2 (4.0.2): do not edit by hand
\name{proxfun}
\alias{proxfun}
\title{Node proximity}
\usage{
proxfun(graph, method, v1 = NULL, v2 = v1, ...)
}
\arguments{
\item{graph}{an object of class igraph}

\item{method}{a method (single string) for calculating similarities, see Details}

\item{v1,v2}{vectors of vertices between which similarity will be calculated
character is treated as names, numeric as ids}

\item{...}{additional arguments specific for a selected method}
}
\value{
matrix
}
\description{
General function for calculating nodes' proximity in a graph
}
\details{
This function calculates vertex proximity (similarity) with selected method
  and between selected vertices.

Available methods are:
\itemize{
  \item \code{act} average commute time
  \item \code{act_n} normalized average commute time
  \item \code{aa} Adamic-Adar index
  \item \code{cn} common neighbours
  \item \code{cos} cosine similarity
  \item ...
  }
}

