% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_link_predictors.R
\name{prune_recover}
\alias{prune_recover}
\title{Performance evaluation of link predictors}
\usage{
prune_recover(g, ..., probes = seq(0.1, 0.5, 0.1), epochs = 10,
  metric = "recall@k", preserve_conn = FALSE, use_weights = FALSE)
}
\arguments{
\item{g}{igraph; The network of interest.}

\item{...}{character; One or more function names able to produce a tibble
with 3 columns, the first two being node IDs (candidate link) and the third
their likelihood of interaction. This tibble should be sorted from most to 
less likely candidate links.}

\item{probes}{numeric; A vector indicating the fraction of links to be pruned
from \code{g} at random.}

\item{epochs}{integer; Number of times each fraction of links is randomly
removed.}

\item{metric}{character; The metric to evaluate the link predictor(s) 
performance. Should be one of 'recall@k', 'aupr', 'auroc' or 'avg_prec'.}

\item{preserve_conn}{logical; Should network connectivity be preserved after 
edge pruning? This is important for some prediction techniques that can only 
be applied to connected topologies. To preserve connectivity, the network's 
Minimum Spanning Tree (MST) is computed and links are pruned around it.}

\item{use_weights}{logical; If connectivity is to be preserved, this 
parameter indicates whether the MST should be computed taking edge weights 
into account.}
}
\value{
Tibble with the following columns:
\item{method}{Link prediction method.}
\item{epoch}{Epoch of link removal.}
\item{frac_rem}{Fraction of links pruned at the specified epoch.}
\item{links_rem}{Number of links pruned at the specified epoch.}
\item{perf}{The performance of the link predictor, according to the selected
evaluation metric.}
}
\description{
Given a network of interest and any number of link prediction function names,
assesses the performance of these link predictors by pruning edges from the 
network and checking if the methods give high likelihood scores to the
removed edges.
}
\examples{
# Assess the performance of three link predictors applied to the Zachary 
# Karate Club network
assessment <- prune_recover(g = karate_club, "lp_cn", "lp_aa", "lp_pa")

}
\references{
Lu, L. et al. (2015) Toward link predictability of complex 
networks. \emph{PNAS} 112(8):2325-2330
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
